--- src/main/java/com/sparta/realtomatoapp/auth/config/SecurityConfig.java
@@ -1,5 +1,0 @@
-package com.sparta.realtomatoapp.auth.config;
-
-public class SecurityConfig {
-    //인증 및 인가 규칙을 구성, 복붙 후 커스터마이징 필요
-}
