+++ src/main/java/com/sparta/realtomatoapp/common/entity/BaseAuditingEntity.java
@@ -1,8 +1,7 @@
-package com.sparta.realtomatoapp.common;
+package com.sparta.realtomatoapp.common.entity;
 
 import jakarta.persistence.*;
 import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDateTime;
