--- src/main/java/com/sparta/realtomatoapp/config/WebConfig.java
@@ -1,5 +1,0 @@
-package com.sparta.realtomatoapp.config;
-
-public class WebConfig {
-    // 혹시 몰라서 남겨둬봄
-}
