--- src/main/java/com/sparta/realtomatoapp/domain/menu/controller/1
@@ -1,1 +1,0 @@
-1
